requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bleach==3.3.0        
build==0.5.1
certifi==2021.5.30   
chardet==4.0.0       
check-manifest==0.46 
colorama==0.4.4      
cycler==0.10.0
distlib==0.3.2
docutils==0.17.1
fake-useragent==0.1.11
filelock==3.0.12
idna==2.10
importlib-metadata==4.6.1
keyring==23.0.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
numpy==1.21.0
packaging==21.0
pandas==1.3.0
pep517==0.10.0
Pillow==8.3.1
pkginfo==1.7.0
psycopg2==2.9.1
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tqdm==4.61.2
twine==3.4.1
urllib3==1.26.6
virtualenv==20.4.7
webencodings==0.5.1
zipp==3.5.0